+++ Sources/GrAIdient/Metal/Kernel/VQSeqHalf.metal
@@ -44,7 +44,7 @@ kernel void vqSeqForwardHalf(
     }
     
     int minIndex = -1;
-    float minValue = 0.0;
+    half minValue = 0.0;
     for (uint k=0; k<K; k++)
     {
         float value = 0.0;
@@ -131,7 +131,7 @@ kernel void vqSeqBackwardHalf(
         uint offsetWeights = depth + nbNeurons * minIndex;
         
         float vq = weights[offsetWeights];
-        float deltaCur = delta[offset];
+        half deltaCur = delta[offset];
         float outPrev = outsPrev[offset];
         
         if (dirty)
